pager almost working

+++ app/src/main/java/com/example/homework2/data/model/Post.kt
@@ -8,6 +8,6 @@ data class Post(
     val id: String,
     val text: String?,
     val images: List<ApiImage>,
-    val dateCreated: Int?,
+    val dateCreated: Long?,
     val likes: ApiLike,
 )
