+++ app/src/main/java/com/example/homework2/data/remote/model/ApiPost.kt
@@ -8,7 +8,7 @@ data class ApiPost(
     val id: String,
     val text: String?,
     val images : List<ApiImage>,
-    val dateCreated: Int?,
+    val dateCreated: Long?,
     val likes : ApiLike,
 
 )
